fix: use "command -v" to find interpreter in $PATH (#3150)
In some environments, `which` doesn't work correctly under Bazel, while
`command -v` does.
I think the difference is that `command` is a shell builtin (and POSIX
compliant), whereas `which` is not:

```
$ sh -c 'builtin command -v python3'
/usr/bin/python3

$ sh -c 'builtin which python3'
sh: line 1: builtin: which: not a shell builtin
```

While `command -v` performs fewer checks under the hood, it is more
portable.