fix(core): do not assume rules_python runtime (#3134)
This change reverts the behaviour where we assume that particular
attributes will be always present - if bazel is doing autoloading for
WORKSPACE builds (7.6.1), then we will crash with attribute error.

I could not think how to add a unit test, which would test this fix
because it seems to only happen with a released version of rules_python
where we are not using `local_repository` override.

Fixes #3119

---------

Co-authored-by: Richard Levasseur <richardlev@gmail.com>