fix(pypi): pull fewer wheels with experimental_index_url (#3058)
Before this we would pull all of the wheels that the user target
configuration
would be compatible with and that meant that it was not customizable.
This also
meant that there were a lot of footguns in the configuration where the
select
statements were not really foolproof.

With this PR we select only those sources that need to be for the
declared
configurations.

Freethreaded support should be done by defining extra freethreaded
platforms
using the new builder API. It is done as a followup in #3063.

This is also changing the default platforms to be only the fully
supported
platforms. This makes the testing easier and avoids us running into
compatibility issues during the rollout.

Work towards #2747
Fixes #2759
Fixes #2849