feat: freethreaded support for the builder API (#3063)
This is a continuation of #3058 where we define freethreaded platforms.
They
need to be used only for particular python versions so I included an
extra
marker configuration attribute where we are using pipstar marker
evaluation
before using the platform.

I think this in general will be a useful tool to configure only
particular
platforms for particular python versions

Fixes #2548, since this shows how we can define custom platforms
Work towards #2747