refactor(pypi): split out a hub_builder helper from the extension code (#3243)
This is a somewhat tedious refactor, where I am just moving code around
(and sometimes renaming various parameters).

I am not modifying and/or fixing any bugs other than more error messages
in one place since I noticed there was a lack of validation. The main
idea is to create a `hub_builder` so that we could also use it for
`pip.configure` calls and/or use it for `py.lock` file parsing and reuse
code. I hope that moving it to a separate file makes it a little bit
more obvious what pieces are used to create a hub repository.

What is more, since the pip extension is reproducible, I have removed
some code that was sorting the output.

Work towards #2747

---------

Co-authored-by: Richard Levasseur <richardlev@gmail.com>