fix: ensure the stage1 bootstrap is executable (#3258)
Bazel tends to make files executable, even if ctx.actions.write() or
ctx.actions.expand_template() is called without is_executable = True.
However, in an analysis tool of mine that is a bit more pedantic than
Bazel this leads to the issue that py_binary() targets can't be executed
due to them not having a +x bit.

Considering that the stage2 bootstrap is marked executable, let's mark
is_executable for consistency.