tests: add non-blocking ci config for bazel rolling (#3272)
With some core Bazel changes to flags coming that will affect us, and
the difficulty it
is to keep the bazel-at-head-and-downstream pipeline green, I figured
it'd be a good idea
to add a CI job that uses the weekly Bazel release so we can identify
problems sooner
and more obviously.

The CI job only run on Ubuntu to save CI slots and is won't block merges
if it has failures.