chore: use python.defaults to set rules_python default python version (#3301)
python.defaults is the modern way to set the default.

Setting it this way also helps avoid a bug where if a root module has a
single `python.toolchain()` call (which are implicitly treated as
`is_default=True`) and also sets the default using `python.defaults()`,
some validation logic gives an error about using both ways to set a
default.