fix(uv): make uv lock rule work with platform python runtime (#3303)
A platform runtime is when the interpreter is provided as an absolute
path to Python
instead of bundled in the runfiles. The uv lock rule almost worked, it
just didn't
handle the runtime files being empty in such a case.

To fix, use an empty depset to satisfy the subsequent APIs that use it.