fix: handle urls without release id format (#3306)
Sometimes when users customize the URLs for release, they use urls that
don't have
the same format the python-build-standalone URLs. Namely, they may not
have
the release_id component of the url. This would result in an error
parsing such
urls.

To fix, refine the url parsing to check if a component is a valid
numeric release id.

Fixes https://github.com/bazel-contrib/rules_python/issues/3285