tests: add tests to verify toolchain registration (#3313)
This adds tests that verify toolchains are registered and resolving
correctly for
the different variants and platforms for the runtimes.

This also shows that workspace mode isn't registering musl or
freethreaded builds
correctly, so use them isn't as easy as simply setting the build flags.
For now,
the tests skip those in workspace mode.

Along the way, fix a bug where py_runtime would crash if it got the
python version
from the flag, and the version contained more than micro (e.g.
"3.14.0rc0")