chore: reject py2 runtimes and remove usages of ctx.fragments.py.disable_py2 (#3319)
This removes the usages of the disable_py2 fragment attribute. This also
makes py2
runtimes rejected (both when set on py_runtime and PyRuntimeInfo)

Because Python 2 support was dropped long ago and other parts of the
code reject,
ignore, or don't work with Python 2 already, this is not considered a
breaking change.

Work towards: https://github.com/bazel-contrib/rules_python/issues/3252

---------

Co-authored-by: Richard Levasseur <rlevasseur@google.com>